### 2017-06-07 #282 @whedon
[REVIEW]: Kindel: indel-aware consensus for nucleotide sequence alignments
**Submitting author:** @bede (<a href="http://orcid.org/0000-0002-3480-3819">Bede Constantinides</a>)
**Repository:** <a href="https://github.com/bede/kindel" target ="_blank">https://github.com/bede/kindel</a>
**Version:** v0.2.0
