### 2017-06-20 #282 @genomematt
Looks quite good.  Only issues are:
The lack of a clear example usage.  I would recommend something that shows some input and output.  (Probably easiest to do by just working through an analysis in an ipynb)

The paper is quite sparse.  It definitely needs to cite ococo like the main page does.  A simplified diagram example of input and output would help a lot, especially for someone slightly off field.

There are guidelines for reporting bugs and support, but no guidelines for contributing.   See other JOSS articles for examples.
