### 2017-06-20 #282 @bede
Thanks for your comments Matt. I'll reply again once I've incorporated the following:
- Example usage (Jupyter notebook or otherwise)
- Additional detail to the paper, perhaps including a diagrammatic explanation of usage
- Contribution guidelines
