### 2017-06-28 2edbb80
Usage notebook, paper, contribution guidelines, figures
