+++ README.md
@@ -3,11 +3,25 @@
 [![Build Status](https://travis-ci.org/bede/kindel.svg?branch=master)](https://travis-ci.org/bede/kindel)  
 [![JOSS status](http://joss.theoj.org/papers/117efd1fc35bb2011311f73d3fa0b545/status.svg)](http://joss.theoj.org/papers/117efd1fc35bb2011311f73d3fa0b545)  
 
-A consensus caller which accepts a headed SAM/BAM input and generates a majority rule consensus in fasta format,  reconciling small indels described in CIGAR fields so as to maximise read-reference concordance. Kindel also permits recovery of consensus sequence across highly divergent regions (such as  those encoding viral envelope proteins) where regions of reads cannot be aligned. With **`--realign`**, Kindel identifies regions of the reference that are clip-dominant (>depth\*0.5) and attempts to assemble a patched consensus using the unaligned sequence context.
+A consensus caller which accepts a headed SAM/BAM input and generates a majority consensus,  reconciling small indels described in CIGAR fields so as to maximise read-reference concordance. Kindel also permits recovery of consensus sequence across highly divergent regions (such as  those encoding viral envelope proteins) where regions of reads cannot be aligned. With **`--realign`**, Kindel identifies regions of the reference that are clip-dominant (>depth\*0.5) and attempts to assemble a patched consensus using the unaligned sequence context. Existing consensus calling approaches are complicated and often involve a variant calling step. An [elegant streaming approach](https://github.com/karel-brinda/ococo) was recently released but cannot reconcile indels.
 
-![clip-dominant region](cdr.png)
 
-Existing consensus calling approaches are complicated and often involve a variant calling step. An [elegant streaming approach](https://github.com/karel-brinda/ococo) was recently released but cannot reconcile indels.
+
+### Core functionality
+
+![clip-dominant region](kindelflow.png)
+
+
+
+
+
+### Reassembly of clip-dominant regions (CDRs) with `--realign`
+
+![clip-dominant region](cdrs.png)
+
+
+
+
 
 
 ## Features
@@ -20,7 +34,10 @@ Existing consensus calling approaches are complicated and often involve a varian
 - [ ] Fast C++ version (early stages)
 
 
+
+
 ## Installation
+
 ```python
 # Requires Python 3.5+
 pip3 install kindel
@@ -28,7 +45,11 @@ pip3 install kindel
 Dependencies should automatically installed, except for Samtools which is needed for BAM input.
 
 
+
 ## Usage
+
+Also see [`usage.ipynb`](usage.ipynb)
+
 ### Command line
 ```
 $ kindel consensus --realign alignment.bam
@@ -127,10 +148,29 @@ kindel.bam_to_consensus(bam_path, realign=False, min_depth=2, min_overlap=7,
 ```
 
 
+
 ## Issues
+
 Please let me know if you run into problems by opening a GitHub issue, [tweeting](https://twitter.com/beconstant) or mailing me via `b at bede dawt im`.
 - Conceived for use with highly diverse Hepatitis C populations – untested with anything larger
+
 - SAM/BAM files must contain an SQ header line containing the reference sequence length
+
 - Able to close gaps of up to 2x read length given adequate depth of coverage
+
 - Sometimes requires multiple runs to converge on an optimal consensus
+
 - Slow (10-20k records/s)
+
+
+
+
+## Contributing
+
+If you would like to contribute to this project, please open an issue or contact the author directly using the details above. Please note that this project is released with a [Contributor Code of Conduct](https://github.com/statsmaths/kerasR/blob/master/CONDUCT.md), and by participating in this project you agree to abide by its terms.
+
+Before issuing a pull request, please:
+
[…]
