### 2017-06-28 #282 @bede
I've done the following:
- Created an example usage notebook
- Added detail to the paper, including two diagrams
- Reviewed a handful of accepted JOSS papers, and added a contribution section to the README

Are you happy with the changes @genomematt ?
