+++ paper.md
@@ -25,7 +25,7 @@ date: 26 May 2017
 
 Kindel is a collection of tools for inferring consensus sequence from an alignment of nucleotide sequences in Sequence Alignment/Map (SAM) format [1]. Sequence substitutions, insertions and deletions (indels) are represented in the generated consensus where the majority of  sequence information—even that contained within unaligned sequence regions—supports their presence. At regions where reads deviate sufficiently from the reference sequence, unaligned sequence context is used to locally reassemble a consensus. In this way, Kindel generates a data-specific reference sequence that maximises overall read-reference similarity. An elegant streaming consensus approach to consensus inference was implemented in OCOCO [2], but like other approaches it fails to reconcile indels.
 
-Kindel was developed for inferring consensus of highly diverse populations of RNA viruses such as hepatitis C and HIV, and is tested with deep sequenced hepatitis C alignments generated by BWA-MEM [3] and Segemehl [4]. Furthermore, Kindel may be used to quantify and visualise subconsensus variation in allele frequencies across a reference sequence, facilitating comparison of intrapatient population state among multiple individuals and/or timepoints. Kindel is implemented as Python 3 package with a command line interface.
+Kindel was developed for inferring consensus of highly diverse populations of RNA viruses such as hepatitis C and HIV, and is tested with deep sequenced hepatitis C alignments generated by BWA-MEM [3] and Segemehl [4]. Furthermore, Kindel may be used to quantify and visualise subconsensus variation in allele frequencies across a reference sequence, facilitating comparison of intrapatient population state among multiple individuals and/or timepoints. Kindel is implemented as a Python 3 package with a command line interface.
 
 
 
