### 2017-07-02 #282 @genomematt
This now ticks all the boxes.  I think we are ready to accept @mgymrek
