### 2017-07-03 #282 @gymreklab
This looks great to me. @arfon I think this is ready to go.
