+++ paper.bib
@@ -1,0 +1,49 @@
+@article{samtools,
+  author = {Li, Heng and Handsaker, Bob and Wysoker, Alec and Fennell, Tim and Ruan, Jue and Homer, Nils and Marth, Gabor and Abecasis, Goncalo and Durbin, Richard and , },
+  title = {The Sequence Alignment/Map format and SAMtools},
+  journal = {Bioinformatics},
+  volume = {25},
+  number = {16},
+  pages = {2078},
+  year = {2009},
+  doi = {10.1093/bioinformatics/btp352},
+  URL = { + http://dx.doi.org/10.1093/bioinformatics/btp352},
+  eprint = {/oup/backfile/content_public/journal/bioinformatics/25/16/10.1093/bioinformatics/btp352/2/btp352.pdf}
+}
+
+@article{ococo,
+  archivePrefix = {arXiv},
+  arxivId = {1605.09070},
+  author = {Břinda, Karel and Boeva, Valentina and Kucherov, Gregory},
+  eprint = {1605.09070},
+  pages = {1-21},
+  title = {{Dynamic read mapping and online consensus calling for better variant detection}},
+  url = {http://arxiv.org/abs/1605.09070},
+  year = {2016}
+}
+
+@article{bwa,
+  author = {Li, Heng and Durbin, Richard},
+  title = {Fast and accurate short read alignment with Burrows–Wheeler transform},
+  journal = {Bioinformatics},
+  volume = {25},
+  number = {14},
+  pages = {1754},
+  year = {2009},
+  doi = {10.1093/bioinformatics/btp324},
+  URL = {http://dx.doi.org/10.1093/bioinformatics/btp324},
+  eprint = {/oup/backfile/content_public/journal/bioinformatics/25/14/10.1093/bioinformatics/btp324/2/btp324.pdf}
+}
+
+@article{segemehl,
+  author = {Otto, Christian and Stadler, Peter F. and Hoffmann, Steve},
+  title = {Lacking alignments? The next-generation sequencing mapper segemehl revisited},
+  journal = {Bioinformatics},
+  volume = {30},
+  number = {13},
+  pages = {1837},
+  year = {2014},
+  doi = {10.1093/bioinformatics/btu146},
+  URL = {http://dx.doi.org/10.1093/bioinformatics/btu146},
+  eprint = {/oup/backfile/content_public/journal/bioinformatics/30/13/10.1093_bioinformatics_btu146/2/btu146.pdf}
+}
