### 2017-07-04 #282 @bede
@arfon Thanks! Now sorted.

I'd seen a lot of JOSS submissions using hardcoded citations and just followed suit. This is much cleaner.
