### 2017-07-05 ff466a5
DataFrame formatting, ConfInts
