### 2017-07-05 #282 @bede
@arfon Also, is it possible to revise the text on this page? http://joss.theoj.org/papers/117efd1fc35bb2011311f73d3fa0b545
