### 2017-07-06 #282 @arfon
> @arfon Also, is it possible to revise the text on this page?

This will be replaced by your paper content once we've accepted the paper.

Thanks for making the edits to `paper.md`, take a look at the compiled PDF here [10.21105.joss.00282.pdf](https://github.com/openjournals/joss-reviews/files/1129066/10.21105.joss.00282.pdf)

I'm wondering if you should remove the sections (`Core functionality` etc) and just replace these with slightly longer captions? The flow of the document in Pandoc is pretty weird. Smaller figures might help too.
