### 2017-07-12 #282 @bede
@arfon Thanks for your comments. I (think I) now understand how Rmd citations and figure captions work. Things should look better next time you compile.
