### 2017-07-13 #282 @arfon
> @arfon Thanks for your comments. I (think I) now understand how Rmd citations and figure captions work. Things should look better next time you compile.

That's looking good. Thanks. At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
