### 2017-07-13 #282 @bede
Thanks – here it is: https://doi.org/10.5281/zenodo.826723
