+++ paper.bib
@@ -1,5 +1,5 @@
 @article{samtools,
-  author = {Li, Heng and Handsaker, Bob and Wysoker, Alec and Fennell, Tim and Ruan, Jue and Homer, Nils and Marth, Gabor and Abecasis, Goncalo and Durbin, Richard and , },
+  author = {Li, Heng and Handsaker, Bob and Wysoker, Alec and Fennell, Tim and Ruan, Jue and Homer, Nils and Marth, Gabor and Abecasis, Goncalo and Durbin, Richard},
   title = {The Sequence Alignment/Map format and SAMtools},
   journal = {Bioinformatics},
   volume = {25},
