### 2017-07-13 #282 @arfon
@bede - your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00282 ⚡️ 🚀 💥
